+++ resolwe/flow/management/__init__.py
@@ -4,9 +4,6 @@
 Flow Management
 ===============
 
-.. automodule:: resolwe.flow.management.commands.purge
-    :members:
-
 .. automodule:: resolwe.flow.management.commands.register
     :members: